+++ e_metrobus/navigation/views.py
@@ -240,13 +240,11 @@ class AnswerView(NavigationView):
         "results": {"enabled": True},
     }
 
-    def get_context_data(self, question, answer=None, **kwargs):
+    def get_context_data(self, question, **kwargs):
         self.title = questions.QUESTIONS[question.category].label
         self.title_icon = questions.QUESTIONS[question.category].small_icon
         context = super(AnswerView, self).get_context_data(**kwargs)
-        context["answer"] = answer
         context["question"] = question
-        context["points"] = questions.SCORE_CORRECT if answer else questions.SCORE_WRONG
         return context
 
     def get(self, request, **kwargs):
@@ -256,6 +254,17 @@ class AnswerView(NavigationView):
         question = questions.get_question_from_name(question_name)
         context = self.get_context_data(question=question, **kwargs)
         return self.render_to_response(context)
+
+
+class AnswerScoreView(TemplateView):
+    template_name = "navigation/answer_score.html"
+
+    def get_context_data(self, question, answer, **kwargs):
+        context = super(AnswerScoreView, self).get_context_data(**kwargs)
+        context["answer"] = answer
+        context["question"] = question
+        context["points"] = questions.SCORE_CORRECT if answer else questions.SCORE_WRONG
+        return context
 
     def post(self, request, **kwargs):
         question = questions.get_question_from_name(request.POST["question"])